pyproject.toml
[tool.poetry]
name = "quotemaker"
version = "0.1.0"
description = ""
authors = ["Calem Roelofs <davidnullsworth@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
spacy = "^3.1.0"
markovify = "^0.9.0"
requests = "^2.25.1"
Pillow = "^8.3.1"

[tool.poetry.dev-dependencies]
black = {version = "^21.6b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
